Move Lockfile and UpdateSpecification to v2prelim
